🍻 refactor: .Recoil Atom 폴더 구조 및 파일 이름 변경
