+++ web/src/state/modal/ModalAtom.ts
@@ -1,5 +1,5 @@
 import { atom } from "recoil";
-import { Todo } from "./todoAtom";
+import { Todo } from "../todo-List/todoAtom";
 
 export const editTodoModalState = atom<{ isOpen: boolean; todo: Todo | null }>({
   key: "editTodoModalState",
